chore(provider): restructure files hierarchy
