+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as Artifacts from './artifacts'
 import { AdminPools, ERC20CollateralPool, Pools } from './pools'
-import * as AssistedProvider from './pools/assisted-provider'
-import * as SelfProvider from './pools/self-provider'
+import * as AssistedProvider from './provider/assisted.provider'
+import * as SelfProvider from './provider/self.provider'
 import * as TypesErc20CollateralToken from './types/erc20-collateral-token'
 import * as TypesPools from './types/pools'
 import * as TypesGeneral from './types/types'
