+++ src/pools/index.ts
@@ -1,5 +1,3 @@
 export * from './admin-pools'
-export * from './assisted-provider'
 export * from './erc20-collateral-pool'
 export * from './pools'
-export * from './self-provider'
