+++ test/integration/self-provider-erc20collateralpool.ts
@@ -2,7 +2,8 @@ import {
   erc20CollateralPoolErrorMessage as ecpErrorMessage,
   poolCommonErrorMessage
 } from '../../src/errors'
-import { ERC20CollateralPool, SelfProvider } from '../../src/pools'
+import { ERC20CollateralPool } from '../../src/pools'
+import { SelfProvider } from '../../src/provider'
 import { Lend, Pool } from '../../src/types/erc20-collateral-token'
 import { Erc20 } from '../../src/utilities/erc20'
 import { sleep } from '../../src/utilities/util'
